Add basic Datetime arithmetic
Datetime now have Add and Sub operator with a Deltatime as the right
hand side of the operation. The feature allows different level of
computation. For example, the Deltatime::seconds level will compute
with leap seconds in mind, although Deltatime::days will ignore them.
Hence, Deltatime::seconds(86400) is not always equal to
Deltatime::days(1), depending on the Datetime it applies to.